+++ cypress-tests-v2/cypress/e2e/configs/Payment/Utils.js
@@ -94,15 +94,3 @@ export function defaultErrorHandler(response, response_data) {
     expect(response.body.error).to.include(response_data.body.error);
   }
 }
-
-export function isoTimeTomorrow() {
-  const now = new Date();
-
-  // Create a new date object for tomorrow
-  const tomorrow = new Date(now);
-  tomorrow.setDate(now.getDate() + 1);
-
-  // Convert to ISO string format
-  const isoStringTomorrow = tomorrow.toISOString();
-  return isoStringTomorrow;
-}
